services/checker/requirements.txt
kafka-python
python-dotenv
python-dateutil